+++ client/src/components/util/AuthHelper.js
@@ -1,3 +1,8 @@
+//
+// LibreTexts Conductor
+// AuthHelper.js
+//
+
 import Cookies from 'js-cookie';
 
 const AuthHelper = {
@@ -6,6 +11,13 @@ const AuthHelper = {
           return true;
         }
         return false;
+    },
+    logout: function() {
+        if (process.env.NODE_ENV === 'production') {
+            Cookies.remove('conductor_access', { path: '/', domain: '.libretexts.org', secure: false });
+        } else {
+            Cookies.remove('conductor_access', { path: '/', domain: 'localhost', secure: false });
+        }
     }
 };